+++ CHANGELOG.md
@@ -7,6 +7,7 @@ ChangeLog
 
 - **New:** Added `sortBy` method.
 - **New:** Added `switchableLabel` column option.
+- **Update** Added support for `class` attribute in toolbar buttons.
 
 #### Extensions